+++ crates/burn-std/src/data/tensor/base.rs
@@ -1,12 +1,8 @@
 use core::f32;
 
-use crate::indexing::AsIndex;
-use alloc::boxed::Box;
 use alloc::format;
-use alloc::vec;
 use alloc::vec::Vec;
-use bytemuck::{AnyBitPattern, CheckedBitPattern, Zeroable, cast_mut, checked::CheckedCastError};
-use core::ops::{Index, IndexMut};
+use bytemuck::checked::CheckedCastError;
 use rand::Rng;
 use thiserror::Error;
 
@@ -14,7 +10,6 @@ use crate::Scalar;
 use crate::distribution::Distribution;
 use crate::element::{Element, ElementConversion};
 use crate::tensor::DType;
-use crate::tensor::ravel_index;
 use crate::{
     AccessError, BoolStore, Bytes, ExecutionError, QuantLevel, QuantMode, QuantScheme, QuantValue,
     QuantizedBytes, Reader, Shape, Writer, bf16, f16,
@@ -88,7 +83,7 @@ pub enum TensorReadError {
 
 impl DataError {
     /// Creates a [`DataError::DTypeMismatch`] for the requested element type `E`.
-    fn dtype_mismatch_as<E: Element>(actual: DType) -> Self {
+    pub(super) fn dtype_mismatch_as<E: Element>(actual: DType) -> Self {
         Self::DTypeMismatch {
             expected: E::dtype(),
             actual,
@@ -195,118 +190,6 @@ impl TensorData {
         );
     }
 
-    /// Returns an [`Index`] view wrapper of the [`TensorData`].
-    ///
-    /// # Example
-    /// ```rust,no_run
-    /// use burn_std::*;
-    ///
-    /// let data = TensorData::from([[1.0, 2.0], [3.0, 4.0]]);
-    /// let shape = data.shape.clone();
-    /// let view: TensorDataView<f64> = data.try_view().unwrap();
-    ///
-    /// assert_eq!(view[&[0, 0]], 1.0);
-    /// assert_eq!(view[&[0, 1]], 2.0);
-    /// assert_eq!(view[&[1, 0]], 3.0);
-    /// assert_eq!(view[&[1, 1]], 4.0);
-    /// ```
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if storage access fails or the dtype, byte representation, or element
-    /// count is incompatible with the requested view.
-    pub fn try_view<E: Element>(&self) -> Result<TensorDataView<'_, E>, DataError> {
-        TensorDataView::<E>::try_view(self)
-    }
-
-    /// Returns a [`TensorDataView<E>`] of the [`TensorData`].
-    ///
-    /// # Example
-    /// ```rust,no_run
-    /// use burn_std::*;
-    ///
-    /// let data = TensorData::from([[1.0, 2.0], [3.0, 4.0]]);
-    /// let shape = data.shape.clone();
-    /// let view: TensorDataView<f64> = data.view();
-    ///
-    /// assert_eq!(view[&[0, 0]], 1.0);
-    /// assert_eq!(view[&[0, 1]], 2.0);
-    /// assert_eq!(view[&[1, 0]], 3.0);
-    /// assert_eq!(view[&[1, 1]], 4.0);
-    /// ```
-    ///
-    /// # Returns
-    /// The view.
-    ///
-    /// # Panics
-    ///
-    /// Panics if the view can't be created because storage access fails or the dtype, byte
-    /// representation, or element count is incompatible with `E`.
-    #[track_caller]
-    pub fn view<E: Element>(&self) -> TensorDataView<'_, E> {
-        self.try_view()
-            .unwrap_or_else(|err| panic!("Failed to create TensorData view: {err}"))
-    }
-
-    /// Returns a [`TensorDataViewMut<E>`] of the [`TensorData`].
-    ///
-    /// # Example
-    /// ```rust,no_run
-    /// use burn_std::*;
-    ///
-    /// let mut data = TensorData::from([[1.0, 2.0], [3.0, 4.0]]);
-    /// let shape = data.shape.clone();
-    /// let mut view: TensorDataViewMut<f64> = data.try_mut_view().unwrap();
-    ///
-    /// assert_eq!(view[&[0, 0]], 1.0);
-    /// assert_eq!(view[&[0, 1]], 2.0);
-    /// assert_eq!(view[&[1, 0]], 3.0);
-    /// assert_eq!(view[&[1, 1]], 4.0);
-    ///
-    /// view[&[0, 0]] = 10.0;
-    /// assert_eq!(view[&[0, 0]], 10.0);
-    /// ```
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if storage access fails or the dtype, byte representation, or element
-    /// count is incompatible with the requested view.
-    pub fn try_mut_view<E: Element>(&mut self) -> Result<TensorDataViewMut<'_, E>, DataError> {
-        TensorDataViewMut::<E>::try_mut_view(self)
-    }
-
-    /// Returns a [`TensorDataViewMut<E>`] of the [`TensorData`].
-    ///
-    /// # Example
-    /// ```rust,no_run
-    /// use burn_std::*;
-    ///
-    /// let mut data = TensorData::from([[1.0, 2.0], [3.0, 4.0]]);
-    /// let shape = data.shape.clone();
-    /// let mut view: TensorDataViewMut<f64> = data.mut_view();
-    ///
-    /// assert_eq!(view[&[0, 0]], 1.0);
-    /// assert_eq!(view[&[0, 1]], 2.0);
-    /// assert_eq!(view[&[1, 0]], 3.0);
-    /// assert_eq!(view[&[1, 1]], 4.0);
-    ///
-    /// view[&[0, 0]] = 10.0;
-    /// assert_eq!(view[&[0, 0]], 10.0);
-    /// ```
-    ///
-    /// # Returns
-    /// The mut view.
-    ///
-    /// # Panics
-    ///
-    /// Panics if the view can't be created because storage access fails or the dtype, byte
-    /// representation, or element count is incompatible with `E`.
-    #[track_caller]
-    pub fn mut_view<E: Element>(&mut self) -> TensorDataViewMut<'_, E> {
-        self.try_mut_view()
-            .unwrap_or_else(|err| panic!("Failed to create mutable TensorData view: {err}"))
-    }
-
     /// Returns the immutable slice view of the tensor data.
     ///
     /// This materializes lazy storage into host-accessible memory when necessary.
@@ -341,94 +224,7 @@ impl TensorData {
         }
     }
 
-    /// Copies and converts the data to a [`Vec<E>`].
-    ///
-    /// By contract, this is equivalent to:
-    /// `data.clone().try_into_vec_as::<E>()`
-    ///
-    /// Particular conversions may provide more efficient implementations.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if storage access fails, the conversion isn't supported, or the stored
-    /// representation or element count is invalid.
-    pub fn try_to_vec_as<E: Element>(&self) -> Result<Vec<E>, DataError> {
-        self.clone().try_into_vec_as::<E>()
-    }
-
-    /// Converts the data to a [`Vec<E>`].
-    ///
-    /// By contract, this is equivalent to:
-    /// `data.try_cast_as::<E>()?.try_into_vec::<E>()`
-    ///
-    /// Particular conversions may provide more efficient implementations.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if storage access fails, the conversion isn't supported, or the stored
-    /// representation or element count is invalid.
-    pub fn try_into_vec_as<E: Element>(self) -> Result<Vec<E>, DataError> {
-        self.try_cast_as::<E>()?.into_vec_unchecked::<E>()
-    }
-
-    /// Copies the stored values to a vector without dtype conversion.
-    #[deprecated(since = "0.22.0", note = "use try_to_vec::<E>()")]
-    pub fn to_vec<E: Element>(&self) -> Result<Vec<E>, DataError> {
-        self.try_to_vec::<E>()
-    }
-
-    /// Converts the stored values into a vector without dtype conversion.
-    #[deprecated(since = "0.22.0", note = "use try_into_vec::<E>()")]
-    pub fn into_vec<E: Element>(self) -> Result<Vec<E>, DataError> {
-        self.try_into_vec::<E>()
-    }
-
-    /// Copies the stored values to a vector without dtype conversion.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if storage access fails, the stored dtype doesn't match `E`, or the byte
-    /// representation is invalid for `E`.
-    pub fn try_to_vec<E: Element>(&self) -> Result<Vec<E>, DataError> {
-        Ok(self.as_slice()?.to_vec())
-    }
-
-    /// Converts the stored values into a vector without dtype conversion.
-    ///
-    /// This may reuse the underlying allocation when its layout and ownership permit it.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if storage access fails, the stored dtype doesn't match `E`, or the byte
-    /// representation is invalid for `E`.
-    pub fn try_into_vec<E: Element>(self) -> Result<Vec<E>, DataError> {
-        // This means we cannot call `into_vec` for QFloat
-        if !self.matches_target_dtype::<E>() {
-            return Err(DataError::dtype_mismatch_as::<E>(self.dtype));
-        }
-
-        self.into_vec_unchecked()
-    }
-
-    /// Returns the tensor data as a vector of scalar values. Does not check dtype.
-    fn into_vec_unchecked<E: Element>(self) -> Result<Vec<E>, DataError> {
-        let mut me = self;
-        me.bytes.read(Reader::new())?;
-        me.bytes = match me.bytes.try_into_vec::<E>() {
-            Ok(elems) => return Ok(elems),
-            Err(bytes) => bytes,
-        };
-
-        // The bytes might have been deserialized and allocated with a different align.
-        // In that case, we have to memcopy the data into a new vector, more suitably allocated
-        Ok(
-            bytemuck::checked::try_cast_slice(me.bytes.read(Reader::new())?)
-                .map_err(DataError::InvalidRepresentation)?
-                .to_vec(),
-        )
-    }
-
-    fn matches_target_dtype<E: Element>(&self) -> bool {
+    pub(super) fn matches_target_dtype<E: Element>(&self) -> bool {
         let target_dtype = E::dtype();
         match self.dtype {
             DType::Bool(BoolStore::U8) => {
@@ -438,127 +234,6 @@ impl TensorData {
                 matches!(target_dtype, DType::U32 | DType::Bool(BoolStore::U32))
             }
             dtype => dtype == target_dtype,
-        }
-    }
-
-    /// Returns an iterator over the values of the tensor data.
-    pub fn iter<E: Element>(&self) -> Box<dyn Iterator<Item = E> + '_> {
-        if E::dtype() == self.dtype {
-            Box::new(bytemuck::checked::cast_slice(&self.bytes).iter().copied())
-        } else {
-            match self.dtype {
-                DType::I8 => Box::new(
-                    bytemuck::checked::cast_slice(&self.bytes)
-                        .iter()
-                        .map(|e: &i8| e.elem::<E>()),
-                ),
-                DType::I16 => Box::new(
-                    bytemuck::checked::cast_slice(&self.bytes)
-                        .iter()
-                        .map(|e: &i16| e.elem::<E>()),
-                ),
-                DType::I32 => Box::new(
-                    bytemuck::checked::cast_slice(&self.bytes)
-                        .iter()
-                        .map(|e: &i32| e.elem::<E>()),
-                ),
-                DType::I64 => Box::new(
-                    bytemuck::checked::cast_slice(&self.bytes)
-                        .iter()
-                        .map(|e: &i64| e.elem::<E>()),
-                ),
-                DType::U8 => Box::new(self.bytes.iter().map(|e| e.elem::<E>())),
-                DType::U16 => Box::new(
-                    bytemuck::checked::cast_slice(&self.bytes)
-                        .iter()
-                        .map(|e: &u16| e.elem::<E>()),
-                ),
-                DType::U32 => Box::new(
-                    bytemuck::checked::cast_slice(&self.bytes)
-                        .iter()
-                        .map(|e: &u32| e.elem::<E>()),
-                ),
-                DType::U64 => Box::new(
-                    bytemuck::checked::cast_slice(&self.bytes)
-                        .iter()
-                        .map(|e: &u64| e.elem::<E>()),
-                ),
-                DType::BF16 => Box::new(
-                    bytemuck::checked::cast_slice(&self.bytes)
-                        .iter()
-                        .map(|e: &bf16| e.elem::<E>()),
-                ),
-                DType::F16 => Box::new(
-                    bytemuck::checked::cast_slice(&self.bytes)
-                        .iter()
-                        .map(|e: &f16| e.elem::<E>()),
-                ),
-                DType::F32 | DType::Flex32 => Box::new(
-                    bytemuck::checked::cast_slice(&self.bytes)
-                        .iter()
-                        .map(|e: &f32| e.elem::<E>()),
-                ),
-                DType::F64 => Box::new(
-                    bytemuck::checked::cast_slice(&self.bytes)
-                        .iter()
-                        .map(|e: &f64| e.elem::<E>()),
-                ),
-                // bool is a byte value equal to either 0 or 1
-                DType::Bool(BoolStore::Native) | DType::Bool(BoolStore::U8) => {
-                    Box::new(self.bytes.iter().map(|e| e.elem::<E>()))
-                }
-                DType::Bool(BoolStore::U32) => Box::new(
-                    bytemuck::checked::cast_slice(&self.bytes)
-                        .iter()
-                        .map(|e: &u32| e.elem::<E>()),
-                ),
-                DType::QFloat(scheme) => match scheme {
-                    QuantScheme {
-                        level: QuantLevel::Tensor | QuantLevel::Block(_),
-                        mode: QuantMode::Symmetric,
-                        value:
-                            QuantValue::Q8F
-                            | QuantValue::Q8S
-                            // Represent sub-byte values as i8
-                            | QuantValue::Q4F
-                            | QuantValue::Q4S
-                            | QuantValue::Q2F
-                            | QuantValue::Q2S,
-                        ..
-                    } => {
-                        // Quantized int8 values
-                        let q_bytes = QuantizedBytes {
-                            bytes: self.bytes.clone(),
-                            scheme,
-                            num_elements: self.num_elements(),
-                        };
-                        let (values, _) = q_bytes.into_vec_i8();
-
-                        Box::new(
-                            values
-                                .iter()
-                                .map(|e: &i8| e.elem::<E>())
-                                .collect::<Vec<_>>()
-                                .into_iter(),
-                        )
-                    }
-                    QuantScheme {
-                        level: QuantLevel::Tensor | QuantLevel::Block(_),
-                        mode: QuantMode::Symmetric,
-                        value:
-                            QuantValue::E4M3 | QuantValue::E5M2 | QuantValue::E2M1,
-                        ..
-                    } => {
-                        unimplemented!("Not yet implemented for iteration");
-                    }
-                    QuantScheme {
-                        level: QuantLevel::BlockTensor { .. },
-                        ..
-                    } => {
-                        unimplemented!("two-level quantization is not supported yet")
-                    }
-                },
-            }
         }
     }
 
@@ -659,274 +334,15 @@ impl TensorData {
     }
 
     // Unchecked, used to overwrite the dtype
-    fn into_bool_u8(mut self) -> Self {
+    pub(super) fn into_bool_u8(mut self) -> Self {
         self.dtype = DType::Bool(BoolStore::U8);
         self
     }
 
     // Unchecked, used to overwrite the dtype
-    fn into_bool_u32(mut self) -> Self {
+    pub(super) fn into_bool_u32(mut self) -> Self {
         self.dtype = DType::Bool(BoolStore::U32);
         self
-    }
-
-    /// Converts the data to the dtype represented by `E`.
-    ///
-    /// # Panics
-    ///
-    /// Panics if storage access fails, the conversion isn't supported, or the stored
-    /// representation or element count is invalid.
-    #[track_caller]
-    pub fn convert<E: Element>(self) -> Self {
-        // TODO: deprecate?
-        self.try_cast_as::<E>()
-            .unwrap_or_else(|err| panic!("Failed to convert TensorData: {err}"))
-    }
-
-    /// Converts the data to `dtype`.
-    ///
-    /// # Panics
-    ///
-    /// Panics if storage access fails, the conversion isn't supported, or the stored
-    /// representation or element count is invalid.
-    #[track_caller]
-    pub fn convert_dtype(self, dtype: DType) -> Self {
-        // TODO: deprecate?
-        self.try_cast(dtype)
-            .unwrap_or_else(|err| panic!("Failed to convert TensorData to {dtype:?}: {err}"))
-    }
-
-    /// Converts the data to the dtype represented by `E`.
-    ///
-    /// By contract, this is equivalent to:
-    /// `data.try_cast(E::dtype())`
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if storage access fails, the conversion isn't supported, or the stored
-    /// representation or element count is invalid.
-    pub fn try_cast_as<E: Element>(self) -> Result<TensorData, DataError> {
-        self.try_cast(E::dtype())
-    }
-
-    /// Converts the data to `dtype`.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if storage access fails, the conversion isn't supported, or the stored
-    /// representation or element count is invalid.
-    pub fn try_cast(self, dtype: DType) -> Result<TensorData, DataError> {
-        if dtype == self.dtype {
-            Ok(self)
-        } else if dtype.size() == self.dtype.size()
-            && !matches!(
-                self.dtype,
-                DType::Bool(BoolStore::Native) | DType::QFloat(_)
-            )
-            && !matches!(dtype, DType::Bool(BoolStore::Native) | DType::QFloat(_))
-        {
-            self.try_cast_inplace(dtype)
-        } else {
-            self.try_cast_clone(dtype)
-        }
-    }
-
-    // Self-to-Self casts should be stripped before this point.
-    fn try_cast_inplace(self, dtype: DType) -> Result<TensorData, DataError> {
-        // Convert self.dtype to generic parameter:
-        match self.dtype {
-            DType::F64 => self.try_cast_inplace_from::<f64>(dtype),
-            DType::F32 | DType::Flex32 => self.try_cast_inplace_from::<f32>(dtype),
-            DType::F16 => self.try_cast_inplace_from::<f16>(dtype),
-            DType::BF16 => self.try_cast_inplace_from::<bf16>(dtype),
-            DType::I64 => self.try_cast_inplace_from::<i64>(dtype),
-            DType::I32 => self.try_cast_inplace_from::<i32>(dtype),
-            DType::I16 => self.try_cast_inplace_from::<i16>(dtype),
-            DType::I8 => self.try_cast_inplace_from::<i8>(dtype),
-            DType::U64 => self.try_cast_inplace_from::<u64>(dtype),
-            DType::U32 => self.try_cast_inplace_from::<u32>(dtype),
-            DType::U16 => self.try_cast_inplace_from::<u16>(dtype),
-            DType::U8 => self.try_cast_inplace_from::<u8>(dtype),
-            DType::Bool(BoolStore::U8) => self.try_cast_inplace_from::<u8>(dtype),
-            DType::Bool(BoolStore::U32) => self.try_cast_inplace_from::<u32>(dtype),
-            DType::Bool(BoolStore::Native) | DType::QFloat(_) => Err(DataError::DTypeMismatch {
-                expected: dtype,
-                actual: self.dtype,
-            }),
-        }
-    }
-
-    fn try_cast_inplace_from<Current>(self, dtype: DType) -> Result<TensorData, DataError>
-    where
-        Current: Element + AnyBitPattern,
-    {
-        // Convert target dtype to generic parameter.
-        match dtype {
-            DType::F64 => self.try_convert_inplace::<Current, f64>(),
-            DType::F32 | DType::Flex32 => self.try_convert_inplace::<Current, f32>(),
-            DType::F16 => self.try_convert_inplace::<Current, f16>(),
-            DType::BF16 => self.try_convert_inplace::<Current, bf16>(),
-            DType::I64 => self.try_convert_inplace::<Current, i64>(),
-            DType::I32 => self.try_convert_inplace::<Current, i32>(),
-            DType::I16 => self.try_convert_inplace::<Current, i16>(),
-            DType::I8 => self.try_convert_inplace::<Current, i8>(),
-            DType::U64 => self.try_convert_inplace::<Current, u64>(),
-            DType::U32 => self.try_convert_inplace::<Current, u32>(),
-            DType::U16 => self.try_convert_inplace::<Current, u16>(),
-            DType::U8 => self.try_convert_inplace::<Current, u8>(),
-            DType::Bool(BoolStore::U8) => self
-                .try_convert_inplace_bool::<Current, u8>()
-                .map(TensorData::into_bool_u8),
-            DType::Bool(BoolStore::U32) => self
-                .try_convert_inplace_bool::<Current, u32>()
-                .map(TensorData::into_bool_u32),
-            DType::Bool(BoolStore::Native) | DType::QFloat(_) => Err(DataError::DTypeMismatch {
-                expected: dtype,
-                actual: Current::dtype(),
-            }),
-        }
-    }
-
-    fn try_convert_inplace<Current, Target>(self) -> Result<TensorData, DataError>
-    where
-        Current: Element + AnyBitPattern,
-        Target: Element + AnyBitPattern,
-    {
-        self.try_convert_inplace_with::<Current, Target>(|x| x.elem())
-    }
-
-    fn try_convert_inplace_bool<Current, Target>(self) -> Result<TensorData, DataError>
-    where
-        Current: Element + AnyBitPattern,
-        Target: Element + AnyBitPattern,
-    {
-        self.try_convert_inplace_with::<Current, Target>(|x| x.to_bool().elem())
-    }
-
-    fn try_convert_inplace_with<Current, Target>(
-        mut self,
-        transform: impl Fn(&Current) -> Target,
-    ) -> Result<TensorData, DataError>
-    where
-        Current: Element + AnyBitPattern,
-        Target: Element + AnyBitPattern,
-    {
-        let expected = self.num_elements();
-        let values =
-            bytemuck::checked::try_cast_slice_mut::<_, Current>(self.bytes.write(Writer::new())?)
-                .map_err(DataError::InvalidRepresentation)?;
-        let actual = values.len();
-        if actual != expected {
-            return Err(DataError::ElementCountMismatch { expected, actual });
-        }
-
-        for x in values {
-            let t = transform(x);
-            let x = cast_mut::<_, Target>(x);
-            *x = t;
-        }
-
-        self.dtype = Target::dtype();
-
-        Ok(self)
-    }
-
-    fn try_cast_clone(self, dtype: DType) -> Result<TensorData, DataError> {
-        // Convert self.dtype to generic parameter:
-        match self.dtype {
-            DType::F64 => self.try_cast_clone_from::<f64>(dtype),
-            DType::F32 | DType::Flex32 => self.try_cast_clone_from::<f32>(dtype),
-            DType::F16 => self.try_cast_clone_from::<f16>(dtype),
-            DType::BF16 => self.try_cast_clone_from::<bf16>(dtype),
-            DType::I64 => self.try_cast_clone_from::<i64>(dtype),
-            DType::I32 => self.try_cast_clone_from::<i32>(dtype),
-            DType::I16 => self.try_cast_clone_from::<i16>(dtype),
-            DType::I8 => self.try_cast_clone_from::<i8>(dtype),
-            DType::U64 => self.try_cast_clone_from::<u64>(dtype),
-            DType::U32 => self.try_cast_clone_from::<u32>(dtype),
-            DType::U16 => self.try_cast_clone_from::<u16>(dtype),
-            DType::U8 => self.try_cast_clone_from::<u8>(dtype),
-            DType::Bool(BoolStore::Native) => self.try_cast_clone_from::<bool>(dtype),
-            DType::Bool(BoolStore::U8) => self.try_cast_clone_from::<u8>(dtype),
-            DType::Bool(BoolStore::U32) => self.try_cast_clone_from::<u32>(dtype),
-            DType::QFloat(_) => Err(DataError::UnsupportedConversion {
-                to: dtype,
-                from: self.dtype,
-            }),
-        }
-    }
-
-    fn try_cast_clone_from<Current>(self, dtype: DType) -> Result<TensorData, DataError>
-    where
-        Current: Element + CheckedBitPattern,
-    {
-        // Convert target dtype to generic parameter.
-        match dtype {
-            DType::F64 => self.try_convert_clone::<Current, f64>(),
-            DType::F32 | DType::Flex32 => self.try_convert_clone::<Current, f32>(),
-            DType::F16 => self.try_convert_clone::<Current, f16>(),
-            DType::BF16 => self.try_convert_clone::<Current, bf16>(),
-            DType::I64 => self.try_convert_clone::<Current, i64>(),
-            DType::I32 => self.try_convert_clone::<Current, i32>(),
-            DType::I16 => self.try_convert_clone::<Current, i16>(),
-            DType::I8 => self.try_convert_clone::<Current, i8>(),
-            DType::U64 => self.try_convert_clone::<Current, u64>(),
-            DType::U32 => self.try_convert_clone::<Current, u32>(),
-            DType::U16 => self.try_convert_clone::<Current, u16>(),
-            DType::U8 => self.try_convert_clone::<Current, u8>(),
-            DType::Bool(BoolStore::Native) => self.try_convert_clone::<Current, bool>(),
-            DType::Bool(BoolStore::U8) => self
-                .try_convert_clone_bool::<Current, u8>()
-                .map(TensorData::into_bool_u8),
-            DType::Bool(BoolStore::U32) => self
-                .try_convert_clone_bool::<Current, u32>()
-                .map(TensorData::into_bool_u32),
-            DType::QFloat(_) => Err(DataError::UnsupportedConversion {
-                to: dtype,
-                from: self.dtype,
-            }),
-        }
-    }
-
-    fn try_convert_clone<Current, Target>(self) -> Result<TensorData, DataError>
-    where
-        Current: Element + CheckedBitPattern,
-        Target: Element + Zeroable,
-    {
-        self.try_convert_clone_with::<Current, Target>(|x| x.elem())
-    }
-
-    fn try_convert_clone_bool<Current, Target>(self) -> Result<TensorData, DataError>
-    where
-        Current: Element + CheckedBitPattern,
-        Target: Element + Zeroable,
-    {
-        self.try_convert_clone_with::<Current, Target>(|x| x.to_bool().elem())
-    }
-
-    fn try_convert_clone_with<Current, Target>(
-        self,
-        transform: impl Fn(&Current) -> Target,
-    ) -> Result<TensorData, DataError>
-    where
-        Current: Element + CheckedBitPattern,
-        Target: Element + Zeroable,
-    {
-        let expected = self.num_elements();
-        let values =
-            bytemuck::checked::try_cast_slice::<_, Current>(self.bytes.read(Reader::new())?)
-                .map_err(DataError::InvalidRepresentation)?;
-        let actual = values.len();
-        if actual != expected {
-            return Err(DataError::ElementCountMismatch { expected, actual });
-        }
-        let mut out: Vec<Target> = vec![Zeroable::zeroed(); expected];
-
-        for (value, out) in values.iter().zip(&mut out) {
-            *out = transform(value);
-        }
-
-        Ok(TensorData::new(out, self.shape))
     }
 
     /// Returns the data as a slice of bytes.
@@ -1105,231 +521,16 @@ impl core::fmt::Display for TensorData {
     }
 }
 
-/// Typed [`Index`] view over a [`TensorData`].
-///
-/// Creating a view materializes lazy storage into host-accessible memory when necessary. It does
-/// not perform dtype conversion.
-///
-/// # Example
-/// ```rust,no_run
-/// use burn_std::*;
-///
-/// let data = TensorData::from([[1.0, 2.0], [3.0, 4.0]]);
-/// let view: TensorDataView<f64> = data.view();
-///
-/// assert_eq!(view.shape(), &data.shape);
-/// assert_eq!(&view.dtype(), &data.dtype);
-///
-/// assert_eq!(view[&[0, 0]], 1.0);
-/// assert_eq!(view[&[0, 1]], 2.0);
-/// assert_eq!(view[&[1, 0]], 3.0);
-/// assert_eq!(view[&[1, 1]], 4.0);
-/// ```
-#[derive(Debug)]
-pub struct TensorDataView<'a, E: Element> {
-    values: &'a [E],
-    shape: &'a Shape,
-    dtype: DType,
-}
-
-impl<'a, E: Element> TensorDataView<'a, E> {
-    /// Creates a typed indexed view over `data`.
-    ///
-    /// # Example
-    /// ```rust,no_run
-    /// use burn_std::*;
-    ///
-    /// let data = TensorData::from([[1.0, 2.0], [3.0, 4.0]]);
-    /// let view: TensorDataView<f64> = data.try_view().unwrap();
-    ///
-    /// assert_eq!(view.shape(), &data.shape);
-    /// assert_eq!(&view.dtype(), &data.dtype);
-    ///
-    /// assert_eq!(view[&[0, 0]], 1.0);
-    /// assert_eq!(view[&[0, 1]], 2.0);
-    /// assert_eq!(view[&[1, 0]], 3.0);
-    /// assert_eq!(view[&[1, 1]], 4.0);
-    /// ```
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if storage access fails or the dtype, byte representation, or element
-    /// count is incompatible with `E`.
-    pub fn try_view(data: &'a TensorData) -> Result<TensorDataView<'a, E>, DataError> {
-        let shape = &data.shape;
-        let dtype = data.dtype;
-        let expected = shape.num_elements();
-        let values = data.as_slice::<E>()?;
-        let actual = values.len();
-
-        if actual != expected {
-            return Err(DataError::ElementCountMismatch { expected, actual });
-        }
-
-        Ok(TensorDataView {
-            values,
-            shape,
-            dtype,
-        })
-    }
-
-    /// Returns the shape of the view.
-    pub fn shape(&self) -> &Shape {
-        self.shape
-    }
-
-    /// Returns the dtype of the view.
-    pub fn dtype(&self) -> DType {
-        self.dtype
-    }
-
-    /// Ravels the index via [`ravel_index`] and the view's shape.
-    pub fn ravel_index<I: AsIndex>(&self, index: &[I]) -> usize {
-        ravel_index(index, self.shape)
-    }
-}
-
-impl<'a, I: AsIndex, E: Element> Index<&[I]> for TensorDataView<'a, E> {
-    type Output = E;
-
-    fn index(&self, index: &[I]) -> &Self::Output {
-        let o = self.ravel_index(index);
-        &self.values[o]
-    }
-}
-
-/// Typed mutable [`IndexMut`] view over a [`TensorData`].
-///
-/// Creating a mutable view materializes lazy storage and performs copy-on-write when necessary.
-/// It does not perform dtype conversion.
-///
-/// # Example
-/// ```rust,no_run
-/// use burn_std::*;
-///
-/// let mut data = TensorData::from([[1.0, 2.0], [3.0, 4.0]]);
-/// let shape = data.shape.clone();
-/// let dtype = data.dtype;
-/// let mut view: TensorDataViewMut<f64> = data.mut_view();
-///
-/// assert_eq!(view.shape(), &shape);
-/// assert_eq!(&view.dtype(), &dtype);
-///
-/// assert_eq!(view[&[0, 0]], 1.0);
-/// assert_eq!(view[&[0, 1]], 2.0);
-/// assert_eq!(view[&[1, 0]], 3.0);
-/// assert_eq!(view[&[1, 1]], 4.0);
-///
-/// view[&[0, 0]] = 10.0;
-/// assert_eq!(view[&[0, 0]], 10.0);
-/// ```
-#[derive(Debug)]
-pub struct TensorDataViewMut<'a, E: Element> {
-    values: &'a mut [E],
-    // `as_mut_slice` borrows the entire `TensorData`, so the view can't also retain a reference to
-    // its shape. Keep an owned copy until storage and metadata can be borrowed as disjoint fields.
-    shape: Shape,
-    dtype: DType,
-}
-
-impl<'a, E: Element> TensorDataViewMut<'a, E> {
-    /// Creates a typed mutable indexed view over `data`.
-    ///
-    /// # Example
-    /// ```rust,no_run
-    /// use burn_std::*;
-    ///
-    /// let mut data = TensorData::from([[1.0, 2.0], [3.0, 4.0]]);
-    /// let shape = data.shape.clone();
-    /// let dtype = data.dtype;
-    ///
-    /// let mut view: TensorDataViewMut<f64> =
-    ///     TensorDataViewMut::try_mut_view(&mut data).unwrap();
-    ///
-    /// assert_eq!(view.shape(), &shape);
-    /// assert_eq!(&view.dtype(), &dtype);
-    ///
-    /// assert_eq!(view[&[0, 0]], 1.0);
-    /// assert_eq!(view[&[0, 1]], 2.0);
-    /// assert_eq!(view[&[1, 0]], 3.0);
-    /// assert_eq!(view[&[1, 1]], 4.0);
-    ///
-    /// view[&[0, 0]] = 10.0;
-    /// assert_eq!(view[&[0, 0]], 10.0);
-    /// ```
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if storage access fails or the dtype, byte representation, or element
-    /// count is incompatible with `E`.
-    pub fn try_mut_view(data: &'a mut TensorData) -> Result<TensorDataViewMut<'a, E>, DataError> {
-        let shape = data.shape.clone();
-        let dtype = data.dtype;
-        let expected = shape.num_elements();
-        let values = data.as_mut_slice::<E>()?;
-        let actual = values.len();
-
-        if actual != expected {
-            return Err(DataError::ElementCountMismatch { expected, actual });
-        }
-
-        Ok(TensorDataViewMut {
-            values,
-            shape,
-            dtype,
-        })
-    }
-
-    /// Returns the shape of the view.
-    pub fn shape(&self) -> &Shape {
-        &self.shape
-    }
-
-    /// Returns the dtype of the view.
-    pub fn dtype(&self) -> DType {
-        self.dtype
-    }
-
-    /// Ravels the dims via [`ravel_index`] and the view's shape.
-    pub fn ravel_index<I: AsIndex>(&self, index: &[I]) -> usize {
-        ravel_index(index, &self.shape)
-    }
-}
-
-impl<'a, I, E> Index<&[I]> for TensorDataViewMut<'a, E>
-where
-    I: AsIndex,
-    E: Element,
-{
-    type Output = E;
-
-    fn index(&self, index: &[I]) -> &Self::Output {
-        let o = self.ravel_index::<I>(index);
-        &self.values[o]
-    }
-}
-
-impl<'a, I, E> IndexMut<&[I]> for TensorDataViewMut<'a, E>
-where
-    I: AsIndex,
-    E: Element,
-{
-    fn index_mut(&mut self, index: &[I]) -> &mut Self::Output {
-        let o = self.ravel_index::<I>(index);
-        &mut self.values[o]
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{AccessPolicy, AllocationController, AllocationProperty, shape};
-    use alloc::vec;
-    use core::mem::{MaybeUninit, align_of, size_of};
-    use rand::{
+    use crate::*;
+    use ::rand::{
         SeedableRng,
         rngs::{StdRng, SysRng},
     };
+    use alloc::vec;
+    use core::mem::{MaybeUninit, align_of, size_of};
 
     #[test]
     fn should_have_rank() {
