+++ crates/burn-std/src/data/tensor/mod.rs
@@ -1,3 +1,6 @@
 mod base;
+mod conversion;
+mod view;
 
 pub use base::*;
+pub use view::*;
